Initial work on splitting down the parsing in to smaller chunks
