+++ Source/CSharp/Expressive/Expressive.Standard/Tokenisation/FunctionTokenIdentifier.cs
@@ -1,0 +1,6 @@
+﻿namespace Expressive.Standard.Tokenisation
+{
+    internal class FunctionTokenIdentifier : ITokenIdentifier
+    {
+    }
+}
