+++ Source/CSharp/Expressive/Expressive.Standard/Tokenisation/ITokenIdentifier.cs
@@ -1,0 +1,6 @@
+﻿namespace Expressive.Standard.Tokenisation
+{
+    internal interface ITokenIdentifier
+    {
+    }
+}
