+++ Source/CSharp/Expressive/Expressive/Expressive.csproj
@@ -57,6 +57,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Context.cs" />
     <Compile Include="Exceptions\ExpressiveException.cs" />
     <Compile Include="Exceptions\FunctionNameAlreadyRegisteredException.cs" />
     <Compile Include="Exceptions\MissingParticipantException.cs" />
@@ -190,6 +191,7 @@
     <Compile Include="Operators\Logical\NotOperator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Token.cs" />
+    <Compile Include="Tokenisation\Tokeniser.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\ExpressiveStrongName.snk">
